fix(boss-hp): record the HP we pushed the boss to, not the post-wipe reset
A party wipe makes the boss walk home and broadcast full HP once. Measured
(2026-08-31, 바실루스 entity 23280) that frame lands 1.495s BEFORE the combat
toggle clears, so the report ticks had already written full HP into the live
report by the time the battle closed: a fight pushed to 25.4% was recorded as
354,439,800/354,439,800 = 100%. Across replay-diag, 194 of 601 non-saturated
unfinished battles (32%) look like this, and the replay's own cast-time HP
says those fights really ended at 17-22%.

Freezing at the end toggle cannot fix it — the contamination is already in.
So the record value is now the lowest remaining HP observed while the battle
was open (DataManager.BattleLowMobHp). Damage only decreases, so in a normal
fight that equals the HP right after the last hit (0 on a kill); it differs
only when the boss heals, and there it agrees with the cumulative damage.

Second, unrelated-but-same-symptom strand: Saturate() clamped HP to
int.MaxValue on the way into storage. 델트라스(27억대)·차원핵 and 11 other
codes cross that (144 battles in replay-diag), which pinned their gauge, boss
HP contribution and battle record to 2,147,483,647. The wire was u64 all
along, so the clamp is gone and the data layer is long. The upload payload
carries no raw HP field (only bossHpContribution), so no schema bump.

+++ dotnet/src/WaffleMeter.App.Wpf/OverlayViewModel.cs
@@ -822,7 +822,7 @@ public sealed class OverlayViewModel : INotifyPropertyChanged
     }
 
     /// <summary>Boss HP readout per React targetInfoDisplayMode.</summary>
-    private static string FormatTargetHp(int remain, int max, double pct, string mode) => mode switch
+    private static string FormatTargetHp(long remain, long max, double pct, string mode) => mode switch
     {
         "percent" => $"{pct:F1}%",
         "remain_percent" => $"{MeterFormat.FormatAmount(remain)}  {pct:F1}%",
