+++ dotnet/src/WaffleMeter.Data/GameData.cs
@@ -43,7 +43,7 @@ public sealed class GameData : ICaptureGameData
     public void SaveNickname(int uid, string nickname, bool isExecutor, int server, int jobByte) => _knownUsers.Add(uid);
     public void SaveUserPower(int uid, int power) { }
     public void SaveSummon(int summonId, int ownerId) { }
-    public void SaveMobHp(int instanceId, int hp) { }
+    public void SaveMobHp(int instanceId, long hp) { }
     public void SaveUseBuff(int uid, int skillCode, long buffStart, long buffEnd, long duration, int actorId) { }
     public void RequestOfficialCharacterLookup(int uid) { }
     public void SavePartyRoster(IReadOnlyList<(string Nickname, int Server, int Slot)> members) { }
