+++ dotnet/tests/WaffleMeter.Capture.Tests/MemberProfileParsingTests.cs
@@ -40,7 +40,7 @@ public sealed class MemberProfileParsingTests
         public void SaveNickname(int uid, string nickname, bool isExecutor, int server, int jobByte) { }
         public void SaveUserPower(int uid, int power) { }
         public void SaveSummon(int summonId, int ownerId) { }
-        public void SaveMobHp(int instanceId, int hp) { }
+        public void SaveMobHp(int instanceId, long hp) { }
         public void SaveUseBuff(int uid, int skillCode, long buffStart, long buffEnd, long duration, int actorId) { }
         public void RequestOfficialCharacterLookup(int uid) { }
     }
