+++ dotnet/tests/WaffleMeter.Capture.Tests/OwnNicknameValidationTests.cs
@@ -37,7 +37,7 @@ public sealed class OwnNicknameValidationTests
         public void EndBattle(int target) { }
         public void SaveUserPower(int uid, int power) { }
         public void SaveSummon(int summonId, int ownerId) { }
-        public void SaveMobHp(int instanceId, int hp) { }
+        public void SaveMobHp(int instanceId, long hp) { }
         public void SaveUseBuff(int uid, int skillCode, long buffStart, long buffEnd, long duration, int actorId) { }
         public void RequestOfficialCharacterLookup(int uid) { }
         public void SavePartyRoster(IReadOnlyList<(string Nickname, int Server, int Slot)> members) { }
